+++ .gitignore
@@ -1,2 +1,6 @@
 .env
-.vscode
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
