+++ go.mod
@@ -3,6 +3,7 @@ module zuri.chat/zccore
 go 1.16
 
 require (
+	github.com/go-playground/locales v0.14.0 // indirect
 	github.com/go-playground/validator/v10 v10.9.0 // indirect
 	github.com/gorilla/mux v1.8.0
 	github.com/joho/godotenv v1.3.0
